Created xmipp_image_extensions code files, ImgSize functions renamed and moved to xmipp_image_extension
git-svn-id: file:///home/roberto/Git/svn@10587 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: a835df2fbaa05b548c24440204f34ac5fc45cee3

+++ trunk/xmipp/libraries/data/matrix2d.h
@@ -93,7 +93,7 @@ void svbksb(Matrix2D< double >& u,
  * val = MAT_ELEM(m, 0, 0);
  * @endcode
  */
-#define MAT_ELEM(m,i,j) ((m).mdata[(i)*(m).mdimx+(j)])
+#define MAT_ELEM(m,i,j) ((m).mdata[((size_t)i)*(m).mdimx+((size_t)j)])
 
 /** X dimension of the matrix
  */
